Add files via upload
modified the shell program to take a single optional command-line argument, which is the string to use as a prompt. If no such argument is given, continue to use the string "jsh: " as your prompt. In the special case that the command-line argument is the one-character string "-", does not print any prompt at all.